refactor(business-logic): reduce the wallet-change rule to ledger recalculation
The rule did two things on wallet.added / wallet.removed: it recalculated
ledger aggregates, and it rebuilt journal entries and token lots around
that. Only the first survives.

recalculateAggregates is ledger maintenance that happened to live in an
accounting rule — it persists isIgnored, ignoredReason and every running
total back onto the ledger rows and rebuilds the APR periods. Removing the
rule outright would have left adding a wallet with no effect on cost basis
or PnL until some unrelated sync happened, which is a change to the
surviving core rather than to accounting.

The vault branch goes entirely, and that is not a reduction in behaviour:
UniswapV3VaultLedgerService.recalculateAggregates takes no wallet set, the
branch never called it, and everything it did do was journal and lot
maintenance. NFT-only is the accurate description of what this rule does
today, not an oversight.

Class name, rule name, file name and queue are unchanged; only the
directory moves, since rules/accounting/ is going away.
