+++ apps/midcurve-business-logic/src/rules/accounting/uniswapv3/index.ts
@@ -6,4 +6,3 @@
 
 export { UniswapV3PostJournalEntriesRule } from './uniswapv3-post-journal-entries';
 export { UniswapV3ReconcileCostBasisRule } from './uniswapv3-reconcile-cost-basis';
-export { UniswapV3ReevaluateOnWalletChangeRule } from './uniswapv3-reevaluate-on-wallet-change';
