+++ apps/midcurve-business-logic/src/rules/index.ts
@@ -51,8 +51,11 @@ export { RefreshCoingeckoTokensRule } from './refresh-coingecko-tokens';
 export * from './close-orders';
 
 // Accounting rules (double-entry journal system, protocol-specific)
-export { UniswapV3PostJournalEntriesRule, UniswapV3ReconcileCostBasisRule, UniswapV3ReevaluateOnWalletChangeRule } from './accounting/uniswapv3';
+export { UniswapV3PostJournalEntriesRule, UniswapV3ReconcileCostBasisRule } from './accounting/uniswapv3';
 export { UniswapV3VaultPostJournalEntriesRule } from './accounting/uniswapv3-vault';
+
+// Position rules (protocol-specific)
+export { UniswapV3ReevaluateOnWalletChangeRule } from './positions/uniswapv3';
 
 // Automation rules
 export { RefuelOperatorRule } from './automation/refuel-operator';
