+++ apps/midcurve-business-logic/src/rules/positions/uniswapv3/index.ts
@@ -1,0 +1,1 @@
+export { UniswapV3ReevaluateOnWalletChangeRule } from './uniswapv3-reevaluate-on-wallet-change';
